Extract a base adapter with sensible defaults
The default behavior is to display the data as a string on the index and
show pages, and render a text field on the form page.
