+++ lib/adapters/string_adapter.rb
@@ -1,19 +1,4 @@
-class StringAdapter
-  def initialize(data)
-    @data = data
-  end
+require_relative "base_adapter"
 
-  def render_show
-    data
-  end
-
-  def render_index
-    data
-  end
-
-  def render_edit(form, attribute_name)
-    form.text_field(attribute_name)
-  end
-
-  attr_reader :data
+class StringAdapter < BaseAdapter
 end
